Add files via upload
The Simon game is the exciting electronic game of lights and sounds in which players must repeat random sequences of lights by pressing the colored pads in the correct order. It's fast-paced play, with lights and sounds that can challenge you. This was recreated in this project using HTML, CSS, and Javascript(jQuery).